Cleaning

+++ test/models.js
@@ -1,7 +1,9 @@
 var fs = require("fs");
 var dl = require("../lib/DataLayer");
 var dc = require("../lib/DataContext");
-fs.unlinkSync("test.db");
+if (fs.existsSync("test.db")) {
+    fs.unlinkSync("test.db");
+}
 var sqlite3 = new dl.Sqlite3DataLayer("test.db");
 var dataCtx = new dc.DataContext(sqlite3);
 var base = dataCtx.createDataModel({
